refactors coauthors component to use latest standards
