+++ src/components/co-authors/README.md
@@ -1,0 +1,2 @@
+# CoAuthors
+Main component for handling selection of coauthors for the Gutenberg editor.
