+++ src/components/co-authors/style.css
@@ -1,0 +1,19 @@
+.cap-combobox {
+	margin-top: 16px;
+}
+
+.cap-icon-button-stack {
+	display: flex;
+	flex-wrap: wrap;
+	width: 30px;
+}
+
+.cap-icon-button-stack .cap-icon-button {
+	padding-bottom: 0;
+	position: relative;
+	height: 20px;
+}
+
+.cap-icon-button-stack .cap-icon-button:nth-child(2n) {
+	padding-top: 0;
+}
